+++ tools/static-assets/server/mini-files.ts
@@ -1,5 +1,4 @@
-import * as assert from "assert";
-import * as path from "path";
+import path from "path";
 import { release, EOL } from "os";
 
 // All of these functions are attached to files.js for the tool;
@@ -18,7 +17,7 @@ export function isWindowsLikeFilesystem() {
   return process.platform === "win32" || release().indexOf("Microsoft") >= 0;
 }
 
-export function toPosixPath(p, partialPath) {
+export function toPosixPath(p: string, partialPath: boolean = false) {
   // Sometimes, you can have a path like \Users\IEUser on windows, and this
   // actually means you want C:\Users\IEUser
   if (p[0] === "\\" && (! partialPath)) {
@@ -36,7 +35,7 @@ export function toPosixPath(p, partialPath) {
 
 export const convertToPosixPath = toPosixPath;
 
-export function toDosPath(p, partialPath) {
+export function toDosPath(p: string, partialPath: boolean = false) {
   if (p[0] === '/' && ! partialPath) {
     if (! /^\/[A-Za-z](\/|$)/.test(p))
       throw new Error("Surprising path: " + p);
@@ -51,27 +50,25 @@ export function toDosPath(p, partialPath) {
 
 export const convertToWindowsPath = toDosPath;
 
-export function convertToOSPath(standardPath, partialPath) {
+export function convertToOSPath(standardPath: string, partialPath: boolean = false) {
   if (process.platform === "win32") {
     return toDosPath(standardPath, partialPath);
   }
-
   return standardPath;
 }
 
-export function convertToStandardPath(osPath, partialPath) {
+export function convertToStandardPath(osPath: string, partialPath: boolean = false) {
   if (process.platform === "win32") {
     return toPosixPath(osPath, partialPath);
   }
-
   return osPath;
 }
 
-export function convertToOSLineEndings(fileContents) {
+export function convertToOSLineEndings(fileContents: string) {
   return fileContents.replace(/\n/g, EOL);
 }
 
-export function convertToStandardLineEndings(fileContents) {
+export function convertToStandardLineEndings(fileContents: string) {
   // Convert all kinds of end-of-line chars to linuxy "\n".
   return fileContents.replace(new RegExp("\r\n", "g"), "\n")
                      .replace(new RegExp("\r", "g"), "\n");
@@ -80,43 +77,45 @@ export function convertToStandardLineEndings(fileContents) {
 
 // Return the Unicode Normalization Form of the passed in path string, using
 // "Normalization Form Canonical Composition"
-export function unicodeNormalizePath(path) {
+export function unicodeNormalizePath(path: string) {
   return (path) ? path.normalize('NFC') : path;
 }
 
 // wrappings for path functions that always run as they were on unix (using
 // forward slashes)
-export function wrapPathFunction(name, partialPaths) {
-  const f = path[name];
-  assert.strictEqual(typeof f, "function");
-
-  return function () {
+export function wrapPathFunction<
+  TArgs extends any[],
+  TResult,
+  F extends (...args: TArgs) => TResult,
+>(
+  f: F,
+  partialPath: boolean = false,
+): F {
+  return function wrapper() {
     if (process.platform === 'win32') {
-      const args = Array.prototype.map.call(
+      const result = f.apply(path, Array.prototype.map.call(
         arguments,
         // if partialPaths is turned on (for path.join mostly)
         // forget about conversion of absolute paths for Windows
-        p => toDosPath(p, partialPaths),
-      );
-      const result = f.apply(path, args);
-      if (typeof result === "string") {
-        return toPosixPath(result, partialPaths);
-      }
-      return result;
-    }
+        p => toDosPath(p, partialPath),
+      ) as TArgs);
 
-    return f.apply(path, arguments);
-  };
+      return typeof result === "string"
+        ? toPosixPath(result, partialPath)
+        : result;
+    }
+    return f.apply(path, arguments as any);
+  } as F;
 }
 
-export const pathJoin = wrapPathFunction("join", true);
-export const pathNormalize = wrapPathFunction("normalize");
-export const pathRelative = wrapPathFunction("relative");
-export const pathResolve = wrapPathFunction("resolve");
-export const pathDirname = wrapPathFunction("dirname");
-export const pathBasename = wrapPathFunction("basename");
-export const pathExtname = wrapPathFunction("extname");
-export const pathIsAbsolute = wrapPathFunction("isAbsolute");
+export const pathJoin = wrapPathFunction(path.join, true);
+export const pathNormalize = wrapPathFunction(path.normalize);
+export const pathRelative = wrapPathFunction(path.relative);
+export const pathResolve = wrapPathFunction(path.resolve);
+export const pathDirname = wrapPathFunction(path.dirname);
+export const pathBasename = wrapPathFunction(path.basename);
+export const pathExtname = wrapPathFunction(path.extname);
+export const pathIsAbsolute = wrapPathFunction(path.isAbsolute);
 export const pathSep = '/';
 export const pathDelimiter = ':';
 export const pathOsDelimiter = path.delimiter;