+++ tools/tests/static-html.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var selftest = require('../tool-testing/selftest.js');
-var files = require('../fs/files.js');
+var files = require('../fs/files');
 import { getUrl } from '../utils/http-helpers.js';
 import { sleepMs } from '../utils/utils.js';