+++ tools/tests/wipe-all-packages.js
@@ -1,6 +1,6 @@
 var selftest = require('../tool-testing/selftest.js');
 var Sandbox = selftest.Sandbox;
-var files = require('../fs/files.js');
+var files = require('../fs/files');
 var utils = require('../utils/utils.js');
 var archinfo = require('../utils/archinfo.js');
 var _ = require('underscore');